@@ -12,7 +12,7 @@ description = ""
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "~3.10"
+python = "~3.11"
 pandas = "~2"
 numpy = "~1"
 plotly = "~5"
