@@ -12,7 +12,7 @@ authors = [
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "~3.10"
+python = "~3.11"
 uvicorn = { extras = ["standard"], version = "^0.23.2" }
 flojoy = { path = "pkgs/flojoy", develop = true }
 fastapi = "^0.103.2"
@@ -34,7 +34,7 @@ pyusb = "^1.2.1"
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.2"
 pytest-cov = "^4.1.0"
-ruff = "^0.1.2"
+ruff = "^0.1.5"
 
 [build-system]
 requires = ["poetry-core"]
