@@ -1,13 +1,13 @@
-import { writeFileSync as FsWriteFileSync } from "fs";
-import api from "../electron/api";
+/* eslint-disable no-var */
+import { BrowserWindow } from "electron";
+import api from "./api/index";
+import { InterpretersList } from "./main/python/interpreter";
 
 declare global {
+  var mainWindow: BrowserWindow;
+  var pythonInterpreters: InterpretersList;
+
   interface Window {
-    api: typeof api & {
-      isPackaged: boolean;
-    };
-    electronAPI: {
-      writeFileSync: typeof FsWriteFileSync;
-    };
+    api: typeof api;
   }
 }
